+++ src/components/issues/CardIssue.css
@@ -12,3 +12,9 @@
   transition: all 400ms;
   cursor: pointer;
 }
+
+.pos-loading-issues {
+  position: absolute;
+  top: 50px;
+  left: 45%;
+}
