+++ src/components/issues/CardIssue.jsx
@@ -6,7 +6,7 @@ import Card from 'react-bootstrap/Card';
 import "./CardIssue.css"
 
 function CardIssue({ data }) {
-  const theme_params = useSelector((state) => state.theme_params);
+  const theme_params = useSelector((state) => state.params.theme_params);
   const { theme } = theme_params
   const [showModal, setShowModal] = useState(false)
   const { image } = data
