+++ src/components/loading/Loading.jsx
@@ -1,11 +1,9 @@
 import React, { useState, useEffect } from 'react';
-import { useSelector } from "react-redux";
 import Spinner from 'react-bootstrap/Spinner';
 import "./Loading.css"
 
-const Loading = ({ data }) => {
+const Loading = ({ data, state, timeWait = 2500 }) => {
   let [error, setError] = useState(false);
-  let loading_state = useSelector((state) => state.loading);
 
   useEffect(() => {
     if (data.length > 0) {
@@ -13,16 +11,16 @@ const Loading = ({ data }) => {
     } else {
       setTimeout(() => {
         setError(true)
-      }, 2000);
+      }, timeWait);
     }
-  }, [data, error])
+  }, [data, error, timeWait])
 
 
   if (error) return (<p className='notFound'>Comics not found</p >)
   return (
     <>
       {
-        loading_state && <Spinner className='spinner' animation="border" variant="danger" />
+        state && <Spinner className='spinner' animation="border" variant="danger" />
       }
     </>
   );
