+++ src/components/login/Profile.jsx
@@ -1,7 +1,7 @@
 import React, { useEffect } from 'react'
 import { useAuth0 } from '@auth0/auth0-react'
 import { useDispatch } from 'react-redux'
-import { creategame } from '../../redux/actions'
+import { creategame } from '../../redux/actions/filters'
