+++ src/index.js
@@ -7,22 +7,22 @@ import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
 import { AuthContextProvider } from './context/authContext'
-import { Elements } from '@stripe/react-stripe-js';
-import { loadStripe } from '@stripe/stripe-js';
+// import { Elements } from '@stripe/react-stripe-js';
+// import { loadStripe } from '@stripe/stripe-js';
 
 
-const stripePromise = loadStripe("pk_test_51LfUl2GKiZGzVPFiXL66alIPbNJN4wqY9ZLCjUk9Uht9CDx2OQ1cvLZznwSIYlWeBUNRkMQNIXqIm6huk4JqOnIb00UB9E1Ck8")
+// const stripePromise = loadStripe("pk_test_51LfUl2GKiZGzVPFiXL66alIPbNJN4wqY9ZLCjUk9Uht9CDx2OQ1cvLZznwSIYlWeBUNRkMQNIXqIm6huk4JqOnIb00UB9E1Ck8")
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
   <React.StrictMode >
     <AuthContextProvider>
       <Provider store={store}>
-        <Elements stripe={stripePromise}>
+        {/* <Elements stripe={stripePromise}> */}
           <BrowserRouter>
             <App />
           </BrowserRouter>
-        </Elements>
+        {/* </Elements> */}
       </Provider>
     </AuthContextProvider>
   </React.StrictMode>
