+++ src/pages/dashboard/dashboardNav/DashboardNav.css
@@ -17,8 +17,7 @@
   overflow-x: hidden;
 }
 
-.navAdm,
-.unselected {
+.navAdm {
   position: fixed;
   width: 300px;
   height: 100%;
@@ -49,9 +48,26 @@
 
 }
 
-.active {
+.dashboard-active {
   background-color: var(--white);
   color: var(--blue) !important;
+  border-top-left-radius: 30px;
+  border-bottom-left-radius: 30px;
+}
+
+.dashboard-inactive {
+  width: 300px;
+  position: fixed;
+  background: var(--blue);
+  border-left: 10px solid var(--blue);
+  transition: 0.5;
+  overflow: hidden;
+}
+
+.dashboard-inactive:hover {
+
+  background: var(--white);
+  color: var(--blue);
   border-top-left-radius: 30px;
   border-bottom-left-radius: 30px;
 }
@@ -124,13 +140,4 @@
   border-radius: 50%;
   box-shadow: 90px -90px 0 10px var(--white);
   pointer-events: none;
-}
-
-
-.selected {
-  background-color: aqua;
-}
-
-.unselected {
-  background-color: red;
 }
