+++ src/pages/dashboard/users/UserCard.css
@@ -1,5 +1,5 @@
 .user-card {
-  width: 90%;
+  min-width: 80%;
   height: 80px;
   border: none;
   border-radius: 20px;
@@ -19,4 +19,65 @@
 .col-item {
   display: flex;
   flex-direction: column;
+}
+
+.switch {
+  position: relative;
+  display: inline-block;
+  width: 45px;
+  height: 24px;
+}
+
+.switch input {
+  opacity: 0;
+  width: 0;
+  height: 0;
+}
+
+.slider {
+  position: absolute;
+  cursor: pointer;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  background-color: #ccc;
+  -webkit-transition: .4s;
+  transition: .4s;
+}
+
+.slider:before {
+  position: absolute;
+  content: "";
+  height: 16px;
+  width: 16px;
+  left: 4px;
+  bottom: 4px;
+  background-color: white;
+  -webkit-transition: .4s;
+  transition: .4s;
+}
+
+input:checked+.slider {
+  background-color: #41a6de;
+}
+
+input:focus+.slider {
+  box-shadow: 0 0 1px #41a6de;
+}
+
+input:checked+.slider:before {
+  -webkit-transform: translateX(20px);
+  -ms-transform: translateX(20px);
+  transform: translateX(20px);
+}
+
+/* Rounded sliders */
+.slider.round {
+  width: 45px;
+  border-radius: 34px;
+}
+
+.slider.round:before {
+  border-radius: 50%;
 }
