+++ src/redux/actions/comics.js
@@ -1,0 +1,83 @@
+import axios from "axios";
+
+const backendURL = process.env.REACT_APP_API;
+
+export const getAllVolumes = () => {
+  // const token = JSON.parse(localStorage.getItem("token"))
+  return async (dispatch) => {
+    const volumes = await axios({
+      method: 'GET',
+      url: `${backendURL}/comics`,
+      // headers: {
+      // "Authorization": `Bearer ${token.token}`
+      // }
+    })
+    return dispatch({
+      type: "GET_ALL_COMICS",
+      payload: volumes.data
+    })
+  }
+}
+
+export const volumeDetail = (id) => {
+  return async (dispatch) => {
+    const volume = await axios({
+      method: 'get',
+      url: `${backendURL}/comics/${String(id)}`,
+    })
+    return dispatch({
+      type: "GET_COMIC",
+      payload: volume.data
+    })
+  }
+}
+
+export const getIssues = (id) => {
+  return async (dispatch) => {
+    const issues = await axios({
+      method: 'get',
+      url: `${backendURL}/comics/issues/${id}`,
+    })
+
+    return dispatch({
+      type: "GET_ISSUES",
+      payload: issues.data
+    })
+  }
+}
+
+export const searchComic = (volume_name) => {
+  return async (dispatch) => {
+    const comics = await axios({
+      method: 'get',
+      url: `${backendURL}/comics/search?name=${volume_name}`
+    })
+    return dispatch({
+      type: "SEARCH_COMICS",
+      payload: comics.data
+    })
+  }
+}
+
+export const addComic = (body) => {
+  return async (dispatch) => {
+    const comic_info = await axios({
+      method: 'post',
+      url: `${backendURL}/comics`,
+      data: body
+    })
+    return dispatch({
+      type: "POST_COMIC",
+      payload: comic_info.data
+    })
+  }
+}
+
+export const reset_comicState = (payload) => {
+  return {
+    type: "RESET_STATE",
+    payload
+  }
+}
+
+
